feat(regression-testing): add performance tests
fix(sheepdog): use cdis-jenkins-lib master and fix google test (#62)

* fix(sheepdog): cdis-jenkins-lib master

* fix(sheepdog): remove console.log

* fix(googleServieAccountTests): add @

* fix(timeout): google retires

feat(query-tests): add performance tests for peregrine

feat(export performance tests): add export performance tests

fix(dcf-test): disable failing test (#65)

squash me

+++ codecept.conf.js
@@ -54,9 +54,8 @@ exports.config = {
   },
   bootstrap: './test_setup.js',
   teardown() {
-    if (seleniumSessionId !== undefined)
-    {
-      // session id is a global var retrieved in the helper
+    // session id is a global var retrieved in the helper
+    if ( typeof seleniumSessionId !== 'undefined' && seleniumSessionId ) {
       console.log(`Killing Selenium session ${seleniumSessionId}`);
       request.del(`http://localhost:4444/wd/hub/session/${seleniumSessionId}`);
     }
